fix(sysproc): remove invalid int64 cast in sys_mapzero xv6 types.h only defines uint64, not int64. Return p->vr.start directly.

+++ kernel/sysproc.c
@@ -234,5 +234,5 @@ sys_mapzero(void)
   p->vr.size   = (int)aligned;
   p->vr.active = 1;
   p->sz       += aligned;
-  return (int64)p->vr.start;
+  return p->vr.start;
 }
